✨ refactor: Imports to Prevent Circular Type Refs (#8423)
